refactor: 칼럼 데이터 조합 부분 객체로 분리
