+++ src/main/java/persistence/sql/ddl/CreateColumn.java
@@ -1,0 +1,112 @@
+package persistence.sql.ddl;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Field;
+
+public class CreateColumn {
+
+    private String name;
+    private String type;
+    private String constraints;
+    private String generatedValue;
+
+    public CreateColumn(Field field) {
+        this.name = extractName(field);
+        this.type = parseType(field.getType());
+        this.constraints = extractConstraints(field);
+        this.generatedValue = extractGeneratedValue(field);
+    }
+
+    public static String makeColumn(Field field) {
+        CreateColumn column = new CreateColumn(field);
+        return column.getColumn();
+    }
+
+    private String getColumn() {
+        return name + type + constraints + generatedValue;
+    }
+
+    /**
+     * @Column의 name이 유효하다면 column명으로 설정하여 반환합니다.
+     * 없을 경우 기존 field명을 반환합니다.
+     */
+    private String extractName(Field field) {
+        String columnName = field.getName();
+
+        if (!isAnnotation(field, Column.class)) {
+            return columnName;
+        }
+
+        if (!"".equals(field.getDeclaredAnnotation(Column.class).name())) {
+            columnName = field.getDeclaredAnnotation(Column.class).name();
+        }
+
+        return columnName;
+    }
+
+    /**
+     * @Column의 제약조건을 추가해줍니다.
+     */
+    private String extractConstraints(Field field) {
+        String nullable = "";
+
+        if (isAnnotation(field, Column.class)
+                && !field.getDeclaredAnnotation(Column.class).nullable()) {
+            nullable += " not null";
+        }
+
+        return nullable;
+    }
+
+    /**
+     * @GeneratedValue가 유효시 조건을 추가해 줍니다.
+     */
+    private String extractGeneratedValue(Field field) {
+        if (!isAnnotation(field, GeneratedValue.class)) {
+            return "";
+        }
+
+        return getGeneratedValue(field.getAnnotation(GeneratedValue.class).strategy());
+    }
+
+    /**
+     * 필드의 자료형을 읽어 칼럼 자료형으로 치환합니다.
+     */
+    private <T> String parseType(Class<T> type) {
+        return " " + getType(type);
+    }
+
+    private <T> String getType(Class<T> type) {
+        switch (type.getSimpleName()) {
+            case "int":
+                return "INT";
+            case "Integer":
+                return "INTEGER";
+            case "Long":
+                return "BIGINT";
+            case "String":
+                return "VARCHAR(255)";
+        }
+
+        throw new IllegalArgumentException("유효하지 않은 자료형입니다.");
+    }
+
+    private String getGeneratedValue(GenerationType generationType) {
+        switch (generationType) {
+            case IDENTITY:
+                return " generated by default as identity";
+        }
+
+        throw new IllegalArgumentException("지원하지 않는 GeneratedValue 입니다.");
+    }
+
+    private <A> boolean isAnnotation(Field field, Class<A> aClass) {
+        Class<? extends Annotation> annotation = aClass.asSubclass(Annotation.class);
+
+        return field.isAnnotationPresent(annotation);
+    }
+}
